Time: 111 ms (37.15%), Space: 108.3 MB (5.94%) - LeetHub

+++ 2461-maximum-sum-of-distinct-subarrays-with-length-k/2461-maximum-sum-of-distinct-subarrays-with-length-k.cpp
@@ -11,7 +11,7 @@ public:
         }
 
         ll temp = 0;
-        unordered_map<int,int> freq;
+        unordered_map<ll,ll> freq;
 
          for(int i = 0; i < k; ++i){
             temp += nums[i];
